Feature/comment

+++ src/app/api/post/delete/route.ts
@@ -1,24 +1,39 @@
 import prisma from "@/lib/prisma";
 import { NextRequest, NextResponse } from "next/server";
 import jwt from "jsonwebtoken";
+import { DeleteSchema } from "@/utils/validation";
+import z from "zod";
 
 export async function POST(req: NextRequest) {
-    const { id } = await req.json();
+    const { userId, id } = await req.json();
 
     const JWT_SECRET = process.env.JWT_SECRET;
 
     try {
         // ログインしているかどうかの判定
         const token = req.cookies.get("auth_token")?.value;
+
+        if (token === null || token === undefined) {
+            return NextResponse.json({ message: "ログインしていない。", success: false }, { status: 500 });
+        }
+
         const data = await jwt.verify(token!, JWT_SECRET!);
 
         if (!data) {
             return NextResponse.json({ message: "ログインしていない。", success: false }, { status: 500 });
         }
 
-        if (id) {
+        // 削除用バリデーションチェック
+        const issue = DeleteSchema.safeParse({ userId, id });
+        if (!issue.success) {
+            const validation = z.flattenError(issue.error);
+            const message = validation.fieldErrors ?? null;
+            return NextResponse.json({ message: message, success: false }, { status: 400 });
+        }
+
+        if (id && userId) {
             await prisma.posts.delete({
-                where: { id: id }
+                where: { id: id, userId: userId }
             });
             return NextResponse.json({ message: "投稿は正常に削除されました。", success: true }, { status: 200 });
         } else {
